back out for now
git-svn-id: https://svn.r-project.org/R/trunk@59167 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/appl/machar.c
@@ -117,7 +117,7 @@ void machar(int *ibeta, int *it, int *irnd, int *ngrd, int *machep, int *negep, 
     {
         b = b + b;
         temp = a + b;
-        itemp = (int)(temp - a);
+        itemp = temp - a;
     } while (itemp == 0);
     *ibeta = itemp;
     beta = *ibeta;
