+++ src/main/datetime.c
@@ -537,7 +537,7 @@ static struct tm *localtime0(const double *tp, const int local, struct tm *ltm)
         res->tm_isdst = -1;
 
         /* Try to fix up timezone differences */
-        diff = (int)(guess_offset(res) / 60);
+        diff = guess_offset(res) / 60;
         shift = res->tm_min + 60 * res->tm_hour;
         res->tm_min -= diff;
         validate_tm(res);
@@ -547,7 +547,7 @@ static struct tm *localtime0(const double *tp, const int local, struct tm *ltm)
             res->tm_yday--;
         if (shift - diff > 24)
             res->tm_yday++;
-        diff2 = (int)(guess_offset(res) / 60);
+        diff2 = guess_offset(res) / 60;
         if (diff2 != diff)
         {
             res->tm_min += (diff - diff2);
@@ -629,13 +629,13 @@ attribute_hidden unsigned int TimeToSeed(void)
     {
         struct timespec tp;
         clock_gettime(CLOCK_REALTIME, &tp);
-        seed = (unsigned int)(((uint_least64_t)tp.tv_nsec << 16) ^ tp.tv_sec);
+        seed = ((uint_least64_t)tp.tv_nsec << 16) ^ tp.tv_sec;
     }
 #elif defined(HAVE_GETTIMEOFDAY)
     {
         struct timeval tv;
         gettimeofday(&tv, NULL);
-        seed = (unsigned int)(((uint_least64_t)tv.tv_usec << 16) ^ tv.tv_sec);
+        seed = ((uint_least64_t)tv.tv_usec << 16) ^ tv.tv_sec;
     }
 #else
     /* C89, so must work */
@@ -879,7 +879,7 @@ SEXP attribute_hidden do_asPOSIXct(SEXP call, SEXP op, SEXP args, SEXP env)
     for (i = 0; i < n; i++)
     {
         double secs = REAL(VECTOR_ELT(x, 0))[i % nlen[0]], fsecs = floor(secs);
-        tm.tm_sec = (int)fsecs;
+        tm.tm_sec = fsecs;
         tm.tm_min = INTEGER(VECTOR_ELT(x, 1))[i % nlen[1]];
         tm.tm_hour = INTEGER(VECTOR_ELT(x, 2))[i % nlen[2]];
         tm.tm_mday = INTEGER(VECTOR_ELT(x, 3))[i % nlen[3]];
@@ -975,7 +975,7 @@ SEXP attribute_hidden do_formatPOSIXlt(SEXP call, SEXP op, SEXP args, SEXP env)
     for (i = 0; i < N; i++)
     {
         double secs = REAL(VECTOR_ELT(x, 0))[i % nlen[0]], fsecs = floor(secs);
-        tm.tm_sec = (int)fsecs;
+        tm.tm_sec = fsecs;
         tm.tm_min = INTEGER(VECTOR_ELT(x, 1))[i % nlen[1]];
         tm.tm_hour = INTEGER(VECTOR_ELT(x, 2))[i % nlen[2]];
         tm.tm_mday = INTEGER(VECTOR_ELT(x, 3))[i % nlen[3]];
@@ -996,7 +996,7 @@ SEXP attribute_hidden do_formatPOSIXlt(SEXP call, SEXP op, SEXP args, SEXP env)
             else
             {
                 const char *q = CHAR(STRING_ELT(sformat, i % m));
-                int n = (int)strlen(q) + 50;
+                int n = strlen(q) + 50;
                 char buf2[n];
 #ifdef Win32
                 /* We want to override Windows' TZ names */
