+++ src/nmath/pnchisq.c
@@ -11,7 +11,7 @@
  *    distribution function. Appl.Statist., 41, 478-482.
 
  *  Other parts
- *  Copyright (C) 2000-2012  The R Core Team
+ *  Copyright (C) 2000-2009  The R Core Team
  *  Copyright (C) 2003-2009  The R Foundation
  */
 
@@ -92,8 +92,7 @@ double attribute_hidden pnchisq_raw(double x, double f, double theta, double err
 
     if (theta < 80)
     { /* use 110 for Inf, as ppois(110, 80/2, lower.tail=FALSE) is 2e-20 */
-        long double sum = 0, sum2 = 0, lambda = 0.5 * theta,
-                    pr = expl(-lambda); // does this need a feature test?
+        long double sum = 0, sum2 = 0, lambda = 0.5 * theta, pr = exp(-lambda);
         double ans;
         int i;
         /* we need to renormalize here: the result could be very close to 1 */
@@ -104,7 +103,7 @@ double attribute_hidden pnchisq_raw(double x, double f, double theta, double err
             if (sum2 >= 1 - 1e-15)
                 break;
         }
-        ans = (double)(sum / sum2);
+        ans = sum / sum2;
         return ans;
     }
 
@@ -170,7 +169,7 @@ double attribute_hidden pnchisq_raw(double x, double f, double theta, double err
     }
     else
     {
-        t = expl(lt);
+        t = exp(lt);
 #ifdef DEBUG_pnch
         REprintf(", t=exp(lt)= %g\n", t);
 #endif
@@ -220,7 +219,7 @@ double attribute_hidden pnchisq_raw(double x, double f, double theta, double err
 #ifdef DEBUG_pnch
                 REprintf(" n=%d; nomore underflow in u = exp(lu) ==> change\n", n);
 #endif
-                v = u = expl(lu); /* the first non-0 'u' */
+                v = u = exp(lu); /* the first non-0 'u' */
                 lamSml = FALSE;
             }
         }
@@ -238,7 +237,7 @@ double attribute_hidden pnchisq_raw(double x, double f, double theta, double err
 #ifdef DEBUG_pnch
                 REprintf("  n=%d; nomore underflow in t = exp(lt) ==> change\n", n);
 #endif
-                t = expl(lt); /* the first non-0 't' */
+                t = exp(lt); /* the first non-0 't' */
                 tSml = FALSE;
             }
         }
@@ -261,5 +260,5 @@ double attribute_hidden pnchisq_raw(double x, double f, double theta, double err
 #ifdef DEBUG_pnch
     REprintf("\n == L_End: n=%d; term= %g; bound=%g\n", n, term, bound);
 #endif
-    return (double)(lower_tail ? ans : 1 - ans);
+    return lower_tail ? ans : 1 - ans;
 }
